# api/openapi.yaml
swagger: '2.0'
info:
  description: API Documentation
  version: '1.0'
  title: MicroserviceManagement API
  termsOfService: urn:tos
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
host: cm-microservice-management.cloud.iai.kit.edu
basePath: "/"
...
paths:
...
  "/microservices":
    get:
      tags:
        - microservice-management-controller
      summary: returns list of registered microservices
      responses:
        '200':
          description: OK
          schema: 
            type: array
            items:
              "$ref": "#/definitions/Microservice"
        '502':
          description: Bad Gateway
  "/microservice/{deployment_name}":
    get:
      tags:
      - microservice-management-controller
      summary: returns data about a specified microservice
      parameters:
        - in: path
          name: deploymentName
          type: string
          required: true
          description: The name of the deployment
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/definitions/Microservice"
        '502':
          description: Bad Gateway
...
definitions:
  HelmDependency:
    type: object
    properties:
      alias:
        type: string
      name:
        type: string
      repository:
        type: string
      version:
        type: string
    title: HelmDependency
...
  Link:
    type: object
    properties:
      name:
        type: string
      value:
        type: string
    title: Link
  Health: 
    type: object
    properties: 
      state: 
        type: string
      condition: 
        type: string
      pods:
        type: array
        items:
          $ref: "#/definitions/Pod"
    title: Health
  Pod:
    type: object
    properties: 
      name:
        type: string
      replicaSetName:
        type: string
      phase:
        type: string
    title: Pod
  Microservice:
    type: object
    properties:
      hostname:
        type: string
      version:
        type: string
      address:
        type: string
      codeOwner:
        type: string
      dependencies:
        type: array
        items:
          "$ref": "#/definitions/HelmDependency"
      links:
        type: array
        items:
          "$ref": "#/definitions/Link"
      health:
        $ref: "#/definitions/Health"
    title: Microservice
...