# src/main/java/edu/kit/microservice_management/logic/model/Link.java
package edu.kit.microservice_management.logic.model;

import java.beans.ConstructorProperties;

public class Link {
    private String name;
    private String value;

    @ConstructorProperties({"name", "value"})
    public Link(String name, String value) {
        this.name = name;
        this.value = value;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getValue() {
        return value;
    }

    public void setValue(String value) {
        this.value = value;
    }


}